Provisioning service - token refresh logic
