+++ provisioning/service/src/Authentication/ProvisioningServiceClientCredentials.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using Microsoft.Azure.Devices.Common;
+using Microsoft.Azure.Devices.Shared;
 using Microsoft.Rest;
 using System.Net;
 using System.Net.Http;
@@ -23,17 +23,19 @@ namespace Microsoft.Azure.Devices.Provisioning.Service
         /// <param name="request"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
-        public override Task ProcessHttpRequestAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+        public override async Task ProcessHttpRequestAsync(HttpRequestMessage request, CancellationToken cancellationToken)
         {
-            request.Headers.Add(HttpRequestHeader.Authorization.ToString(), GetSasToken());
+            string sasToken = await GetSasToken().ConfigureAwait(false);
+
+            request.Headers.Add(HttpRequestHeader.Authorization.ToString(), sasToken);
             request.Headers.Add("User-Agent", _productInfo.ToString());
-            return base.ProcessHttpRequestAsync(request, cancellationToken);
+            await base.ProcessHttpRequestAsync(request, cancellationToken).ConfigureAwait(false);
         }
 
         /// <summary>
         /// Return the SAS token
         /// </summary>
         /// <returns></returns>
-        protected abstract string GetSasToken();
+        public abstract Task<string> GetSasToken();
     }
 }
