+++ provisioning/service/src/Authentication/ServiceConnectionString.cs
@@ -1,56 +1,70 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using Microsoft.Azure.Devices.Common;
 using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Linq.Expressions;
 using System.Text;
+using System.Text.RegularExpressions;
+using SharedAccessSignatureParser = Microsoft.Azure.Devices.Provisioning.Service.SharedAccessSignature;
 
 namespace Microsoft.Azure.Devices.Provisioning.Service
 {
     /// <summary>
-    /// This object handles the connection string for the Azure IoT Services.
+    /// The Service Connection String Parser class
     /// </summary>
-    /// <remarks>
-    /// The connection string contains a set of information that uniquely identify an IoT Service.
-    /// 
-    /// A valid connection string shall be in one of the following formats:
-    /// <code>
-    /// HostName=[ServiceName];SharedAccessKeyName=[keyName];SharedAccessKey=[Key]
-    /// HostName=[ServiceName];SharedAccessKeyName=[keyName];SharedAccessSignature=[Signature]
-    /// </code>
-    /// 
-    /// This object parse and store the connection string. It is responsible to provide the authorization token too. 
-    /// </remarks>
-    public sealed class ServiceConnectionString
+    public class ServiceConnectionString
     {
-        private static readonly TimeSpan DefaultTokenTimeToLive = TimeSpan.FromMinutes(5);
+        private const char ValuePairDelimiter = ';';
+        private const char ValuePairSeparator = '=';
+        private const string HostNameSeparator = ".";
 
-        /// <summary>
-        /// CONSTRUCOR
-        /// </summary>
-        /// <param name="parser">the <see cref="ServiceConnectionStringParser"/> with the connection string content.</param>
-        /// <exception cref="ArgumentNullException">if the provided parser is null.</exception>
-        public ServiceConnectionString(ServiceConnectionStringParser parser)
+        private const string HostNamePropertyName = nameof(HostName);
+        private const string SharedAccessKeyNamePropertyName = nameof(SharedAccessKeyName);
+        private const string SharedAccessKeyPropertyName = nameof(SharedAccessKey);
+        private const string SharedAccessSignaturePropertyName = nameof(SharedAccessSignature);
+
+        private static readonly TimeSpan regexTimeoutMilliseconds = TimeSpan.FromMilliseconds(500);
+        private static readonly Regex HostNameRegex = new Regex(@"[a-zA-Z0-9_\-\.]+$", RegexOptions.Compiled | RegexOptions.IgnoreCase, regexTimeoutMilliseconds);
+        private static readonly Regex SharedAccessKeyNameRegex = new Regex(@"^[a-zA-Z0-9_\-@\.]+$", RegexOptions.Compiled | RegexOptions.IgnoreCase, regexTimeoutMilliseconds);
+        private static readonly Regex SharedAccessKeyRegex = new Regex(@"^.+$", RegexOptions.Compiled | RegexOptions.IgnoreCase, regexTimeoutMilliseconds);
+        private static readonly Regex SharedAccessSignatureRegex = new Regex(@"^.+$", RegexOptions.Compiled | RegexOptions.IgnoreCase, regexTimeoutMilliseconds);
+
+        private string _hostName;
+        private string _serviceName;
+
+        private ServiceConnectionString()
         {
-            if (parser == null)
-            {
-                throw new ArgumentNullException(nameof(parser));
-            }
-
-            HostName = parser.HostName;
-            SharedAccessKeyName = parser.SharedAccessKeyName;
-            SharedAccessKey = parser.SharedAccessKey;
-            SharedAccessSignature = parser.SharedAccessSignature;
-            ServiceName = parser.ServiceName;
-            HttpsEndpoint = new UriBuilder("https", parser.HostName).Uri;
         }
 
         /// <summary>
-        /// The Provisioning Service Name
+        /// Factory for new Connection String object.
         /// </summary>
-        public string ServiceName
+        /// <remarks>
+        /// The connection string contains a set of information that uniquely identify an IoT Service.
+        /// 
+        /// A valid connection string shall be in the following format:
+        /// <code>
+        /// HostName=[ServiceName];SharedAccessKeyName=[keyName];SharedAccessKey=[Key]
+        /// </code>
+        /// 
+        /// This object parse the connection string providing the artifacts to the <see cref="ServiceConnectionString"/> object. 
+        /// </remarks>
+        /// <param name="serviceConnectionString">the <code>string</code> with the connection string information.</param>
+        /// <returns>A <code>ServiceConnectionStringParser</code> object with the parsed connection string.</returns>
+        public static ServiceConnectionString Create(string serviceConnectionString)
         {
-            get;
-            private set;
+            if (string.IsNullOrWhiteSpace(serviceConnectionString))
+            {
+                throw new ArgumentNullException(nameof(serviceConnectionString));
+            }
+
+            var ServiceConnectionStringParser = new ServiceConnectionString();        
+            ServiceConnectionStringParser.Parse(serviceConnectionString);
+
+            return ServiceConnectionStringParser;
         }
 
         /// <summary>
@@ -58,99 +72,164 @@ namespace Microsoft.Azure.Devices.Provisioning.Service
         /// </summary>
         public string HostName
         {
-            get;
-            private set;
-        }
-
-        /// <summary>
-        /// The Provisioning Service Client Https Endpoint
-        /// </summary>
-        public Uri HttpsEndpoint
-        {
-            get;
-            private set;
-        }
-
-        /// <summary>
-        /// The Provisioning Service Audience
-        /// </summary>
-        public string Audience
-        {
-            get { return HostName; }
+            get { return _hostName; }
+            set { SetHostName(value); }
         }
 
         /// <summary>
         /// The Provisioning Service Access Policy Name
         /// </summary>
-        public string SharedAccessKeyName
-        {
-            get;
-            private set;
-        }
+        public string SharedAccessKeyName { get; internal set; }
 
         /// <summary>
         /// The Provisioning Service Shared Access Key for the specified
         /// access policy
         /// </summary>
-        public string SharedAccessKey
-        {
-            get;
-            private set;
-        }
+        public string SharedAccessKey { get; internal set; }
 
         /// <summary>
         /// The Provisioning Service Shared Access Signature
         /// </summary>
-        public string SharedAccessSignature
+        public string SharedAccessSignature { get; internal set; }
+
+        /// <summary>
+        /// The Provisioning Service Name
+        /// </summary>
+        public string ServiceName
         {
-            get;
-            private set;
+            get { return _serviceName; }
         }
 
         /// <summary>
-        /// Returns the shared access signature for authorization
+        /// Returns the Provisioning Service Connection string
         /// </summary>
         /// <returns></returns>
-        public string GetSasToken()
+        public override string ToString()
         {
-            string password;
-            if (string.IsNullOrWhiteSpace(SharedAccessSignature))
+            Validate();
+
+            StringBuilder stringBuilder = new StringBuilder();
+            stringBuilder.AppendKeyValuePairIfNotEmpty(HostNamePropertyName, HostName);
+            stringBuilder.AppendKeyValuePairIfNotEmpty(SharedAccessKeyNamePropertyName, SharedAccessKeyName);
+            stringBuilder.AppendKeyValuePairIfNotEmpty(SharedAccessKeyPropertyName, SharedAccessKey);
+            stringBuilder.AppendKeyValuePairIfNotEmpty(SharedAccessSignaturePropertyName, SharedAccessSignature);
+            if (stringBuilder.Length > 0)
             {
-                TimeSpan timeToLive;
-                password = BuildToken(out timeToLive);
-            }
-            else
-            {
-                password = SharedAccessSignature;
+                stringBuilder.Remove(stringBuilder.Length - 1, 1);
             }
 
-            return password;
+            return stringBuilder.ToString();
         }
 
-        /// <summary>
-        /// Parser for the Provisioning Service Connection String
-        /// </summary>
-        /// <param name="connectionString"> The DPS Connection String </param>
-        /// <returns></returns>
-        public static ServiceConnectionString Parse(string connectionString)
+        private void Parse(string serviceConnectionString)
         {
-            var parser = ServiceConnectionStringParser.Create(connectionString);
-            return new ServiceConnectionString(parser);
+            IDictionary<string, string> map = serviceConnectionString.ToDictionary(ValuePairDelimiter, ValuePairSeparator);
+
+            HostName = GetConnectionStringValue(map, HostNamePropertyName);
+            SharedAccessKeyName = GetConnectionStringOptionalValue(map, SharedAccessKeyNamePropertyName);
+            SharedAccessKey = GetConnectionStringOptionalValue(map, SharedAccessKeyPropertyName);
+            SharedAccessSignature = GetConnectionStringOptionalValue(map, SharedAccessSignaturePropertyName);
+
+            Validate();
         }
 
-        private string BuildToken(out TimeSpan ttl)
+        private void Validate()
         {
-            var builder = new SharedAccessSignatureBuilder()
+            if (string.IsNullOrWhiteSpace(SharedAccessKeyName))
             {
-                KeyName = SharedAccessKeyName,
-                Key = SharedAccessKey,
-                TimeToLive = DefaultTokenTimeToLive,
-                Target = Audience
-            };
+                throw new ArgumentException("Should specify SharedAccessKeyName");
+            }
 
-            ttl = builder.TimeToLive;
+            if (!(string.IsNullOrWhiteSpace(SharedAccessKey) ^ string.IsNullOrWhiteSpace(SharedAccessSignature)))
+            {
+                throw new ArgumentException("Should specify either SharedAccessKey or SharedAccessSignature");
+            }
 
-            return builder.ToSignature();
+            if (string.IsNullOrWhiteSpace(ServiceName))
+            {
+                throw new FormatException("Missing service name");
+            }
+
+            if (!string.IsNullOrWhiteSpace(SharedAccessKey))
+            {
+                Convert.FromBase64String(SharedAccessKey);
+            }
+
+            if (SharedAccessSignatureParser.IsSharedAccessSignature(SharedAccessSignature))
+            {
+                SharedAccessSignatureParser.Parse(ServiceName, SharedAccessSignature);
+            }
+            
+            ValidateFormat(HostName, HostNamePropertyName, HostNameRegex);
+            ValidateFormatIfSpecified(SharedAccessKeyName, SharedAccessKeyNamePropertyName, SharedAccessKeyNameRegex);
+            ValidateFormatIfSpecified(SharedAccessKey, SharedAccessKeyPropertyName, SharedAccessKeyRegex);
+            ValidateFormatIfSpecified(SharedAccessSignature, SharedAccessSignaturePropertyName, SharedAccessSignatureRegex);
+        }
+
+        private void SetHostName(string hostname)
+        {
+            if (string.IsNullOrWhiteSpace(hostname))
+            {
+                throw new ArgumentNullException(nameof(hostname));
+            }
+
+            ValidateFormat(hostname, HostNamePropertyName, HostNameRegex);
+            _hostName = hostname;
+            SetServiceName();
+        }
+
+        private void SetServiceName()
+        {
+            _serviceName = GetServiceName(HostName);
+
+            if (string.IsNullOrWhiteSpace(ServiceName))
+            {
+                throw new FormatException("Missing service name");
+            }
+        }
+
+        private static void ValidateFormat(string value, string propertyName, Regex regex)
+        {
+            if (!regex.IsMatch(value))
+            {
+                throw new ArgumentException(
+                    string.Format(CultureInfo.InvariantCulture, "The connection string has an invalid value for property: {0}", propertyName), nameof(value));
+            }
+        }
+
+        private static void ValidateFormatIfSpecified(string value, string propertyName, Regex regex)
+        {
+            if (!string.IsNullOrEmpty(value))
+            {
+                ValidateFormat(value, propertyName, regex);
+            }
+        }
+
+        private static string GetConnectionStringValue(IDictionary<string, string> map, string propertyName)
+        {
+            string value;
+            if (!map.TryGetValue(propertyName, out value))
+            {
+                throw new ArgumentException(
+                    string.Format(CultureInfo.InvariantCulture, "The connection string is missing the property: {0}", propertyName), 
+                    nameof(map));
+            }
+
+            return value;
+        }
+
+        private static string GetConnectionStringOptionalValue(IDictionary<string, string> map, string propertyName)
+        {
+            string value;
+            map.TryGetValue(propertyName, out value);
+            return value;
+        }
+
+        private static string GetServiceName(string hostName)
+        {
+            int index = hostName.IndexOf(HostNameSeparator, StringComparison.OrdinalIgnoreCase);
+            string serviceName = index >= 0 ? hostName.Substring(0, index) : hostName;
+            return serviceName;
         }
     }
 }
