+++ provisioning/service/src/ProvisioningServiceClientFactory.cs
@@ -1,11 +1,9 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-using Microsoft.Azure.Devices.Provisioning.Service;
+
 using Microsoft.Rest;
 using System;
-using System.Collections.Generic;
 using System.Net.Http;
-using System.Text;
 
 namespace Microsoft.Azure.Devices.Provisioning.Service
 {
@@ -36,19 +34,21 @@ namespace Microsoft.Azure.Devices.Provisioning.Service
         /// <returns></returns>
         public static ProvisioningServiceClient CreateFromConnectionString(String connectionString, HttpClientHandler rootHandler, params DelegatingHandler[] handlers)
         {
-            _provisioningConnectionString = ServiceConnectionString.Parse(connectionString);
-            Uri dpsHost = _provisioningConnectionString.HttpsEndpoint;
+            _provisioningConnectionString = ServiceConnectionString.Create(connectionString);
+
+            Uri dpsHost = new UriBuilder("https", _provisioningConnectionString.HostName).Uri;
             string sharedAccessSignature = _provisioningConnectionString.SharedAccessSignature;
+            ServiceClientCredentials credentials;
+
             if (sharedAccessSignature != null)
             {
-                var credential = new SharedAccessSignatureCredentials(sharedAccessSignature);
-                return new ProvisioningServiceClient(dpsHost, credential, rootHandler, handlers);
+                credentials = new CredentialsWithToken(sharedAccessSignature);
             }
             else
             {
-                var credential = new SharedAccessKeyCredentials(_provisioningConnectionString);
-                return new ProvisioningServiceClient(dpsHost, credential, rootHandler, handlers);
+                credentials = new CredentialsWithSakRefresh(_provisioningConnectionString);
             }
+            return new ProvisioningServiceClient(dpsHost, credentials, rootHandler, handlers);
         }
     }
 }
